agentic-tutor/backend/requirements.txt
# === Core Web & API ===
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1

# === Database & ORM ===
sqlmodel==0.0.18
sqlalchemy==2.0.35
alembic==1.13.2

# === LangChain Ecosystem (Critical for LangGraph + RAG) ===
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
langgraph==0.2.5
langgraph-checkpoint-sqlite==1.0.0

# === LLM & Tools ===
openai==1.47.0
httpx==0.27.2
tavily-python==0.5.0
arxiv==2.1.3
sympy==1.13.3

# === RAG Pipeline ===
sentence-transformers==3.1.1
faiss-cpu==1.8.0.post1
rich==13.8.1

# === Frontend Dashboard ===
streamlit==1.38.0
plotly==5.24.1
pandas==2.2.2
graphviz==0.20.3
python-dotenv==1.0.1

# === Utilities ===
pydantic==2.9.2
requests==2.32.3